+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 v0.8.6
 - [BUGFIX] save budget as double not as integer
 - [TASK] Update Budgetoverview View after setting a new Budget
-- 
+- [REFACTOR] selectAll Return type in the DataController -> return List of Specific class objects
+- [TASK] change budget colors
+- [TASK] add total budget progressbar